Retrieving the final status of a completed program

+++ server/internal/util/program.go
@@ -8,7 +8,7 @@ import (
 
 // Struct to represent the program for the queue
 type Program struct {
-	Id   int       `json:"id"`   // The id of the program
+	Id   int32       `json:"id"`   // The id of the program
 	TimeCreated time.Time `json:"created"` // The time the node was created
 	Prog []uint32  `json:"prog"`  // The value of the node
 	Name string `json:"progName"` // The name of the program
@@ -17,7 +17,7 @@ type Program struct {
 // Function that creates a new program
 func NewProgram(data *RunStruct) *Program {
 	program := Program {
-		Id: rand.Int(),
+		Id: int32(rand.Int()),
 		TimeCreated: time.Now(),
 		Prog: data.Binary,
 	}
